pyproject.toml
[project]
name = "gallery"
version = "0.1.0"
description = "SmartGallery - Intelligent Photo Gallery with AI Tagging"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Django>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "django-storages>=1.14.2",
    "boto3>=1.34.0",
    "Pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "django-allauth>=0.57.0",
    "django-allauth[socialaccount]>=0.57.0",
    "djangorestframework>=3.14.0",
    "celery>=5.3.0",
    "flower>=2.0.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "python3-openid==3.2.0",
    "pyjwt==2.10.1",
    "python3-saml==1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.test_settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--reuse-db",
    "--nomigrations",
    "-v",
]
testpaths = ["app/tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
